ref: Move view description helper to SwiftDescriptor
Rename to getSanitizedViewDescription and host it alongside the other
description helpers. Move its unit tests to SwiftDescriptorTests.
